refactor footer
